move method strong_orientation
