+++ src/sage/graphs/orientations.py
@@ -14,6 +14,7 @@ etc.). It also implements some iterators over all these orientations.
 
     :meth:`orient` | Return an oriented version of `G` according the input function `f`.
     :meth:`acyclic_orientations` | Return an iterator over all acyclic orientations of an undirected graph `G`.
+    :meth:`strong_orientation` | Return a strongly connected orientation of the graph `G`.
     :meth:`strong_orientations_iterator` | Return an iterator over all strong orientations of a graph `G`
     :meth:`random_orientation` | Return a random orientation of a graph `G`
 
@@ -484,6 +485,120 @@ def acyclic_orientations(G):
         D = DiGraph([(u, v) if label else (v, u) for (u, v), label in orientation.items()])
         D.relabel(perm=reverse_vertex_labels, inplace=True)
         yield D
+
+
+def strong_orientation(G):
+    r"""
+    Return a strongly connected orientation of the graph `G`.
+
+    An orientation of an undirected graph is a digraph obtained by giving an
+    unique direction to each of its edges. An orientation is said to be strong
+    if there is a directed path between each pair of vertices. See also the
+    :wikipedia:`Strongly_connected_component`.
+
+    If the graph is 2-edge-connected, a strongly connected orientation can be
+    found in linear time. If the given graph is not 2-connected, the orientation
+    returned will ensure that each 2-connected component has a strongly
+    connected orientation.
+
+    OUTPUT: a digraph representing an orientation of the current graph
+
+    .. NOTE::
+
+        - This method assumes that the input the graph is connected.
+        - The time complexity is `O(n+m)` for ``SparseGraph`` and `O(n^2)` for
+          ``DenseGraph`` .
+
+    .. SEEALSO::
+
+        - :meth:`~sage.graphs.graph.Graph.orientations`
+        - :meth:`~sage.graphs.orientations.strong_orientations_iterator`
+        - :meth:`~sage.graphs.digraph_generators.DiGraphGenerators.nauty_directg`
+        - :meth:`~sage.graphs.orientations.random_orientation`
+
+    EXAMPLES:
+
+    For a 2-regular graph, a strong orientation gives to each vertex an
+    out-degree equal to 1::
+
+        sage: g = graphs.CycleGraph(5)
+        sage: g.strong_orientation().out_degree()
+        [1, 1, 1, 1, 1]
+
+    The Petersen Graph is 2-edge connected. It then has a strongly connected
+    orientation::
+
+        sage: g = graphs.PetersenGraph()
+        sage: o = g.strong_orientation()
+        sage: len(o.strongly_connected_components())
+        1
+
+    The same goes for the CubeGraph in any dimension::
+
+        sage: all(len(graphs.CubeGraph(i).strong_orientation().strongly_connected_components()) == 1
+        ....:     for i in range(2,6))
+        True
+
+    A multigraph also has a strong orientation::
+
+        sage: g = Graph([(1,2),(1,2)], multiedges=True)
+        sage: g.strong_orientation()
+        Multi-digraph on 2 vertices
+    """
+    d = DiGraph(multiedges=G.allows_multiple_edges())
+    i = 0
+
+    # The algorithm works through a depth-first search. Any edge used in the
+    # depth-first search is oriented in the direction in which it has been
+    # used. All the other edges are oriented backward
+
+    v = next(G.vertex_iterator())
+    seen = {}
+    i = 1
+
+    # Time at which the vertices have been discovered
+    seen[v] = i
+
+    # indicates the stack of edges to explore
+    next_ = G.edges_incident(v)
+
+    while next_:
+        e = next_.pop()
+
+        # Ignore loops
+        if e[0] == e[1]:
+            continue
+
+        # We assume e[0] to be a `seen` vertex
+        e = e if seen.get(e[0], False) is not False else (e[1], e[0], e[2])
+
+        # If we discovered a new vertex
+        if seen.get(e[1], False) is False:
+            d.add_edge(e)
+            next_.extend(ee for ee in G.edges_incident(e[1])
+                         if ((e[0], e[1]) != (ee[0], ee[1])) and ((e[0], e[1]) != (ee[1], ee[0])))
+            i += 1
+            seen[e[1]] = i
+
+        # Else, we orient the edges backward
+        else:
+            if seen[e[0]] < seen[e[1]]:
+                d.add_edge(e[1], e[0], e[2])
+            else:
+                d.add_edge(e)
+
+    # Case of multiple edges. If another edge has already been inserted, we add
+    # the new one in the opposite direction.
+    tmp = None
+    for e in G.multiple_edges():
+        if tmp == (e[0], e[1]):
+            if d.has_edge(e[0], e[1]):
+                d.add_edge(e[1], e[0], e[2])
+            else:
+                d.add_edge(e)
+        tmp = (e[0], e[1])
+
+    return d
 
 
 def strong_orientations_iterator(G):
